スコア計算処理を追加、時間経過で注文が変わる処理を追加、操作説明の表示、時間切れでリザルト画面に移行する、正解、不正解および時間切れ時のSEを追加、BGMの追加

+++ DirectXGame/Code/application/scene/GameClearScene.cpp
@@ -41,7 +41,7 @@ void GameClearScene::Update()
 	{
 		if (!titleFlag)
 		{
-			DebugText::GetInstance()->Print(500, 350, 4, "%d", Sushi::score);
+			DebugText::GetInstance()->Print(500, 350, 4, "%d", (int)Sushi::score);
 		}
 		if (seFlag)
 		{
